Add kubic branding
Review make_spec and directory naming so we can generate both the kubic
and the caasp branding package.

Add kubic image files.

Add a explicit provides so dependencies will be satisfied, even if the package
is called kubic-dex-branding, it still needs to
provide caasp-dex-branding as this is the branding for caasp-dex.

If caasp-dex gets renamed,this provides should be removed.

Signed-off-by: Jordi Massaguer Pla <jmassaguerpla@suse.de>